feat: v2 instrumentation and some refactoring
