+++ engine/crates/engine-v2/Cargo.toml
@@ -39,6 +39,7 @@ mini-moka = { version = "0.10", optional = true, features = ["sync"] }
 config = { package = "engine-v2-config", path = "./config" }
 engine-value = { path = "../engine/value" }
 engine-parser = { path = "../engine/parser" }
+grafbase-tracing = { package = "tracing", path = "../../../engine/crates/tracing" }
 schema = { path = "./schema", package = "engine-v2-schema" }
 # needed for errors, not entirely sure if we should be separate or not right now.
 # might move it back to engine, the goal isn't to rewrite everything from engine per Se
